requirements.txt
requests
pillow
irc
pycurl
lxml
